+++ api/src/main/java/net/jqwik/api/constraints/Negative.java
@@ -3,7 +3,7 @@ package net.jqwik.api.constraints;
 import java.lang.annotation.*;
 
 /**
- * Constrain the range of a generated number to be 0 or less.
+ * Constrain the range of a generated number to be less than 0.
  *
  * Applies to numeric parameters which are also annotated with {@code @ForAll}.
  *
@@ -12,13 +12,6 @@ import java.lang.annotation.*;
  */
 @Target({ ElementType.ANNOTATION_TYPE, ElementType.PARAMETER, ElementType.TYPE_USE })
 @Retention(RetentionPolicy.RUNTIME)
-@ShortRange(min = Short.MIN_VALUE, max = -0)
-@ByteRange(min = Byte.MIN_VALUE, max = -0)
-@IntRange(min = Integer.MIN_VALUE, max = -0)
-@LongRange(min = Long.MIN_VALUE, max = -0L)
-@FloatRange(min = -Float.MAX_VALUE, max = -0f)
-@DoubleRange(min = -Double.MAX_VALUE, max = -0d)
-@BigRange(max = "0")
 @Documented
 public @interface Negative {
 }