+++ engine/src/test/java/net/jqwik/api/constraints/PositiveProperties.java
@@ -8,41 +8,41 @@ class PositiveProperties {
 
 	@Property
 	boolean positiveBytes(@ForAll @Positive byte value) {
-		return value >= 0;
+		return value > 0;
 	}
 
 	@Property
 	boolean positiveShorts(@ForAll @Positive short value) {
-		return value >= 0;
+		return value > 0;
 	}
 
 	@Property
 	boolean positiveIntegers(@ForAll @Positive int value) {
-		return value >= 0;
+		return value > 0;
 	}
 
 	@Property
 	boolean positiveLongs(@ForAll @Positive long value) {
-		return value >= 0;
+		return value > 0;
 	}
 
 	@Property
 	boolean positiveFloats(@ForAll @Positive float value) {
-		return value >= 0;
+		return value > 0;
 	}
 
 	@Property
 	boolean positiveDoubles(@ForAll @Positive double value) {
-		return value >= 0;
+		return value > 0;
 	}
 
 	@Property
 	boolean positiveBigDecimals(@ForAll @Positive BigDecimal value) {
-		return value.compareTo(BigDecimal.ZERO) >= 0;
+		return value.compareTo(BigDecimal.ZERO) > 0;
 	}
 
 	@Property
 	boolean positiveBigIntegers(@ForAll @Positive BigInteger value) {
-		return value.compareTo(BigInteger.ZERO) >= 0;
+		return value.compareTo(BigInteger.ZERO) > 0;
 	}
 }